+++ express/blocks/floating-button/floating-button.js
@@ -3,10 +3,12 @@ import {
   createFloatingButton,
   collectFloatingButtonData,
 } from '../shared/floating-cta.js';
+import {
+  formatDynamicCartLink,
+} from '../../scripts/utils/pricing.js';
 
 export default async function decorate(block) {
   addTempWrapper(block, 'floating-button');
-
   if (block.classList.contains('spreadsheet-powered')) {
     const audience = block.querySelector(':scope > div').textContent.trim();
     if (audience === 'mobile') {
@@ -24,6 +26,7 @@ export default async function decorate(block) {
 
     const blockLinks = blockWrapper.querySelectorAll('a');
     if (blockLinks && blockLinks.length > 0) {
+      await formatDynamicCartLink(blockLinks[0]);
       const linksPopulated = new CustomEvent('linkspopulated', {
         detail: blockLinks,
       });
